+++ src/azure-cli-core/azure/cli/core/tests/test_util.py
@@ -10,8 +10,8 @@ import tempfile
 from datetime import date, time, datetime
 
 from azure.cli.core.util import \
-    (get_file_json, todict, to_snake_case, truncate_text, shell_safe_json_parse, b64_to_hex,
-     hash_string, random_string)
+    (get_file_json, todict, to_snake_case, truncate_text, shell_safe_json_parse,
+     safe_yaml_parse, b64_to_hex, hash_string, random_string, CLIError)
 
 
 class TestUtils(unittest.TestCase):
@@ -160,6 +160,31 @@ class TestUtils(unittest.TestCase):
         self.assertEqual(
             len(failed_strings), 0,
             'The following patterns failed: {}'.format(failed_strings))
+
+    def test_safe_yaml_parse(self):
+        valid_yaml = '''---
+list_elements:
+- element1:
+    property: value
+    another_property: value
+- element2: ["one", "two"]
+'''
+        dict_result = safe_yaml_parse(valid_yaml)
+        self.assertEqual(dict_result['list_elements'][0]['element1']['property'], 'value')
+
+        try:
+            safe_yaml_parse('', throw_on_empty=True)
+        except CLIError:
+            pass
+        else:
+            raise ValueError('Parsing empty yaml should have failed')
+
+        try:
+            safe_yaml_parse('[hello world}')
+        except CLIError:
+            pass
+        else:
+            raise ValueError('Parsing invalid yaml should have failed')
 
     def test_hash_string(self):
         def _run_test(length, force_lower):