+++ src/command_modules/azure-cli-resource/HISTORY.rst
@@ -40,6 +40,12 @@ Release History
 +++++++++++++++++++
 * minor fixes
 
+unreleased
+----------
+
+* Add support for YAML template files
+
+
 2.0.10 (2017-07-07)
 +++++++++++++++++++
 * `group deployment create`: Improve prompting for missing parameters. Improve parsing of `--parameters KEY=VALUE` syntax.